requirements.txt
streamlit
folium
matplotlib.pyplot
requests